nmea: handle Unicore UNHEADINGA sentence
This now includes the heading information provided by the Unicore
UNIHEADINGA sentence.

Currently, the heading information is flipped by 180 degrees with the
assumption that the master antenna would be in front and the slave in
the rear as a default configuration, so without offset set.

This means the master antenna would be the same as what is referred to
as "rover antenna".

Signed-off-by: Julian Oes <julian@oes.ch>