ubx: ignore deprecated MON-HW message
It turns out that the MON-HW message is sent with a shorter length with
protocol version 34 (e.g. uBlox M10).

For reference, see the interface description:
https://content.u-blox.com/sites/default/files/u-blox%20M10-SPG-5.00_InterfaceDescription_UBX-20053845.pdf
at page 104 of 184.

This means we don't need to print a warning about a wrong length, and
can just ignore the packet.

The data that we are interested in is also sent in the MON-RF message
and already parsed in the driver.

Signed-off-by: Julian Oes <julian@oes.ch>